refactor: DefaultMoveStrategy를 이용한 중복 코드 제거
